[이인호]연회비패키지 수정기능처리 완료

+++ NationRentalSystem/src/main/java/kr/or/nationRental/annualfeePakage/service/AnnualfeePakageService.java
@@ -108,20 +108,24 @@ public class AnnualfeePakageService {
 		//annualfee_pakage테이블의 데이터 update 처리
 		annualfeePakageDao.updateAnnualfeePakage(annualfeePakageDto);
 		
-		/*//annualfee_pakage_authority테이블의 데이터 delete 처리
+		//annualfee_pakage_authority테이블의 데이터 delete 처리
 		//위에 deleteAnnualfeePakageAuthority는 해당 패키지내에 적용 행정기관을 모두 삭제하는 것이고
 		//이번에 것은 선택된 행정기관만 삭제하는 것이다
+		if(annualfeePakageDto.getAnnualfeePakageAuthorityCode() != null) {
 			for(int i=0; i<annualfeePakageDto.getAnnualfeePakageAuthorityCode().size(); i++) {
-				annualfeePakageDao.deleteCheckAnnualfeePakageAuthority(annualfeePakageDto.getAnnualfeePakageAuthorityCode().get(i));
-			}*/
+					annualfeePakageDao.deleteCheckAnnualfeePakageAuthority(annualfeePakageDto.getAnnualfeePakageAuthorityCode().get(i));
+			}
+		}
 		
 		//annualfee_pakage_authority테이블의 데이터 insert 처리
+		if(annualfeePakageDto.getAdminagencyCode() != null) {
 			for(int i= 0; i<annualfeePakageDto.getAdminagencyCode().size(); i++) {
 				Map<String, Integer> map = new HashMap<String, Integer>();
 				map.put("adminagencyCode", annualfeePakageDto.getAdminagencyCode().get(i));
 				map.put("annualfeePakageCode", annualfeePakageDto.getAnnualfeePakageCode());
 				annualfeePakageDao.insertAnnualfeePakageAuthority(map);				
-			}	
+			}
+		}
 		
 	}
